+++ stdlib/Unicode/test/runtests.jl
@@ -419,13 +419,13 @@ end
 end
 
 @testset "Unicode equivalence" begin
-    @test isequivalent("no\u00EBl", "noe\u0308l")
-    @test !isequivalent("no\u00EBl", "noe\u0308l ")
-    @test isequivalent("", "")
-    @test !isequivalent("", " ")
-    @test !isequivalent("no\u00EBl", "NOËL")
-    @test isequivalent("no\u00EBl", "NOËL", casefold=true)
-    @test !isequivalent("no\u00EBl", "noel")
-    @test isequivalent("no\u00EBl", "noel", stripmark=true)
-    @test isequivalent("no\u00EBl", "NOEL", stripmark=true, casefold=true)
+    @test isequal_normalized("no\u00EBl", "noe\u0308l")
+    @test !isequal_normalized("no\u00EBl", "noe\u0308l ")
+    @test isequal_normalized("", "")
+    @test !isequal_normalized("", " ")
+    @test !isequal_normalized("no\u00EBl", "NOËL")
+    @test isequal_normalized("no\u00EBl", "NOËL", casefold=true)
+    @test !isequal_normalized("no\u00EBl", "noel")
+    @test isequal_normalized("no\u00EBl", "noel", stripmark=true)
+    @test isequal_normalized("no\u00EBl", "NOEL", stripmark=true, casefold=true)
 end